rotate rect

+++ src/asia/sejong/freedrawing/figures/FDElementFigure.java
@@ -6,7 +6,7 @@ import org.eclipse.swt.graphics.RGB;
 import asia.sejong.freedrawing.model.FDElement;
 
 public interface FDElementFigure extends IFigure {
-	void setLineWidthEx(int lineWidth);
+	void setLineWidthEx(float lineWidth);
 	void setLineStyleEx(int lineStyle);
 	void setLineColorEx(RGB rgbColor);
